cloud/gateway/pyproject.toml
[project]
name = "blockly-gateway"
version = "1.0.0"
description = "Blockly小车云端网关服务"
requires-python = ">=3.9"
dependencies = [
    "python-socketio==5.11.0",
    "aiohttp==3.9.1",
    "uvicorn==0.24.0",
    "uvloop==0.19.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
